requirements.txt
discord.py 
aiohttp 
asyncio 
systemd
python-dotenv